refactor: extract PoolMatchList from PoolView
PoolView 1714→1313 lignes. renderMatchListView() déplacé dans
pool/PoolMatchList.tsx (443 lignes) avec React.memo et props typées.
